+++ mec_input.cpp
@@ -371,9 +371,7 @@ void mec_input::DoHarmonize()
 
     oenum_alb_or_anb const alb_anb =
         static_cast<oenum_alb_or_anb>
-            (static_cast<int>
-                (database_->Query(DB_AgeLastOrNearest)
-                )
+            (database_->Query(DB_AgeLastOrNearest)
             );
     DateOfBirth.minimum_and_maximum
         (minimum_birthdate(IssueAge.maximum(), EffectiveDate.value(), alb_anb)
@@ -497,9 +495,7 @@ void mec_input::DoTransmogrify()
 
     oenum_alb_or_anb const alb_anb =
         static_cast<oenum_alb_or_anb>
-            (static_cast<int>
-                (database_->Query(DB_AgeLastOrNearest)
-                )
+            (database_->Query(DB_AgeLastOrNearest)
             );
 
     int apparent_age = attained_age
